@@ -13,7 +13,7 @@
     "require": {
         "php": "^7.1",
         "elasticsearch/elasticsearch": "^5.2",
-        "laravel/framework": "^5.7"
+        "laravel/framework": "^6.0"
     },
     "require-dev": {
       "phpunit/phpunit": "^7.1",
