@@ -3,12 +3,15 @@
 namespace DesignMyNight\Elasticsearch;
 
 use Closure;
-use DesignMyNight\Elasticsearch\Exceptions\BulkInsertQueryException;
+use DesignMyNight\Elasticsearch\Database\Schema\Blueprint;
+use DesignMyNight\Elasticsearch\Database\Schema\ElasticsearchBuilder;
+use DesignMyNight\Elasticsearch\Database\Schema\Grammars\ElasticsearchGrammar;
 use DesignMyNight\Elasticsearch\Exceptions\QueryException;
 use Elasticsearch\ClientBuilder;
 use Illuminate\Database\Connection as BaseConnection;
 use Illuminate\Database\Events\QueryExecuted;
 use Illuminate\Database\Grammar as BaseGrammar;
+use Illuminate\Support\Arr;
 
 class Connection extends BaseConnection
 {
@@ -24,6 +27,24 @@ class Connection extends BaseConnection
     protected $requestTimeout;
 
     /**
+     * Map configuration array keys with ES ClientBuilder setters
+     *
+     * @var array
+     */
+    protected $configMappings = [
+        'sslVerification'    => 'setSSLVerification',
+        'sniffOnStart'       => 'setSniffOnStart',
+        'retries'            => 'setRetries',
+        'httpHandler'        => 'setHandler',
+        'connectionPool'     => 'setConnectionPool',
+        'connectionSelector' => 'setSelector',
+        'serializer'         => 'setSerializer',
+        'connectionFactory'  => 'setConnectionFactory',
+        'endpoint'           => 'setEndpoint',
+        'namespaces'         => 'registerNamespace',
+    ];
+
+    /**
      * Create a new Elasticsearch connection instance.
      *
      * @param array $config
@@ -31,15 +52,13 @@ class Connection extends BaseConnection
     public function __construct(array $config)
     {
         $this->config = $config;
-
-        $this->indexSuffix = isset($config['suffix']) ? $config['suffix'] : '';
+        $this->indexSuffix = $config['suffix'] ?? '';
 
         // Extract the hosts from config
-        $hostsConfig = $config['hosts'] ?? $config['host'];
-        $hosts = explode(',', $hostsConfig);
+        $hosts = explode(',', $config['hosts'] ?? $config['host']);
 
         // You can pass options directly to the client
-        $options = array_get($config, 'options', []);
+        $options = Arr::get($config, 'options', []);
 
         // Create the connection
         $this->connection = $this->createConnection($hosts, $config, $options);
@@ -49,70 +68,157 @@ class Connection extends BaseConnection
     }
 
     /**
-     * Create a new Elasticsearch connection.
+     * Dynamically pass methods to the connection.
      *
-     * @param array  $hosts
-     * @param array  $config
+     * @param string $method
+     * @param array  $parameters
      *
-     * @return \Elasticsearch\Client
+     * @return mixed
      */
-    protected function createConnection($hosts, array $config, array $options)
+    public function __call($method, $parameters)
     {
-        // apply config to each host
-        $hosts = array_map(function($host) use ($config) {
-            $port = !empty($config['port']) ? $config['port'] : 9200;
-
-            $scheme = !empty($config['scheme']) ? $config['scheme'] : 'http';
-
-            // force https for port 443
-            $scheme = (int) $port === 443 ? 'https' : $scheme;
-
-            return [
-                'host'   => $host,
-                'port'   => $port,
-                'scheme' => $scheme,
-                'user'   => !empty($config['username']) ? $config['username'] : null,
-                'pass'   => !empty($config['password']) ? $config['password'] : null,
-            ];
-        }, $hosts);
-
-        return ClientBuilder::create()
-            ->setHosts($hosts)
-            ->setSelector('\Elasticsearch\ConnectionPool\Selectors\StickyRoundRobinSelector')
-            ->build();
+        return call_user_func_array([$this->connection, $method], $parameters);
     }
 
     /**
-     * Get the default query grammar instance.
+     * Run an SQL statement and get the number of rows affected.
      *
-     * @return \Illuminate\Database\Query\Grammars\Grammar
+     * @param string $query
+     * @param array  $bindings
+     *
+     * @return int
      */
-    protected function getDefaultQueryGrammar()
+    public function affectingStatement($query, $bindings = [])
     {
-        return $this->withIndexSuffix(new QueryGrammar);
+        //
     }
 
     /**
-     * Set the table prefix and return the grammar.
+     * Start a new database transaction.
      *
-     * @param  \Illuminate\Database\Grammar  $grammar
-     * @return \Illuminate\Database\Grammar
+     * @return void
      */
-    public function withIndexSuffix(BaseGrammar $grammar)
+    public function beginTransaction()
     {
-        $grammar->setIndexSuffix($this->indexSuffix);
-
-        return $grammar;
+        //
     }
 
     /**
-     * Get the default post processor instance.
+     * Commit the active database transaction.
      *
-     * @return Processor
+     * @return void
      */
-    protected function getDefaultPostProcessor()
+    public function commit()
     {
-        return new QueryProcessor();
+        //
+    }
+
+    /**
+     * @param string $index
+     * @param string $name
+     */
+    public function createAlias(string $index, string $name): void
+    {
+        $this->indices()->putAlias(compact('index', 'name'));
+    }
+
+    /**
+     * @param string $index
+     * @param array  $body
+     */
+    public function createIndex(string $index, array $body): void
+    {
+        $this->indices()->create(compact('index', 'body'));
+    }
+
+    /**
+     * Run a select statement against the database and return a generator.
+     *
+     * @param string $query
+     * @param array  $bindings
+     * @param bool   $useReadPdo
+     *
+     * @return \Generator
+     */
+    public function cursor($query, $bindings = [], $useReadPdo = false)
+    {
+        $scrollTimeout = '30s';
+        $limit = $query['size'] ?? 0;
+
+        $scrollParams = [
+            'scroll' => $scrollTimeout,
+            'size' => 100, // Number of results per shard
+            'index' => $query['index'],
+            'body' => $query['body'],
+        ];
+
+        $results = $this->select($scrollParams);
+
+        $scrollId = $results['_scroll_id'];
+
+        $numResults = count($results['hits']['hits']);
+
+        foreach ($results['hits']['hits'] as $result) {
+            yield $result;
+        }
+
+        if (!$limit || $limit > $numResults) {
+            $limit = $limit ? $limit - $numResults : $limit;
+
+            foreach ($this->scroll($scrollId, $scrollTimeout, $limit) as $result) {
+                yield $result;
+            }
+        }
+    }
+
+    /**
+     * Run a delete statement against the database.
+     *
+     * @param string $query
+     * @param array  $bindings
+     *
+     * @return array
+     */
+    public function delete($query, $bindings = [])
+    {
+        return $this->run(
+            $query,
+            $bindings,
+            Closure::fromCallable([$this->connection, 'deleteByQuery'])
+        );
+    }
+
+    /**
+     * @param string $index
+     */
+    public function dropIndex(string $index): void
+    {
+        $this->indices()->delete(compact('index'));
+    }
+
+    /**
+     * Get the timeout for the entire Elasticsearch request
+     * @return float
+     */
+    public function getRequestTimeout(): float
+    {
+        return $this->requestTimeout;
+    }
+
+    /**
+     * @return ElasticsearchBuilder|\Illuminate\Database\Schema\Builder
+     */
+    public function getSchemaBuilder()
+    {
+        return new ElasticsearchBuilder($this);
+    }
+
+    /**
+     * @return ElasticsearchGrammar|\Illuminate\Database\Schema\Grammars\Grammar
+     */
+    public function getSchemaGrammar()
+    {
+        return new ElasticsearchGrammar();
     }
 
     /**
@@ -126,163 +232,7 @@ class Connection extends BaseConnection
     }
 
     /**
-     * Log a query in the connection's query log.
-     *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @param  float|null  $time
-     * @return void
-     */
-    public function logQuery($query, $bindings, $time = null)
-    {
-        $this->event(new QueryExecuted(json_encode($query), $bindings, $time, $this));
-
-        if ($this->loggingQueries) {
-            $this->queryLog[] = compact('query', 'bindings', 'time');
-        }
-    }
-
-    /**
-     * Set the table prefix in use by the connection.
-     *
-     * @param  string  $prefix
-     * @return void
-     */
-    public function setIndexSuffix($suffix)
-    {
-        $this->indexSuffix = $suffix;
-
-        $this->getQueryGrammar()->setIndexSuffix($suffix);
-    }
-
-    /**
-     * Begin a fluent query against a database table.
-     *
-     * @param  string  $table
-     * @return \Illuminate\Database\Query\Builder
-     */
-    public function table($table)
-    {
-    }
-
-    /**
-     * Get a new raw query expression.
-     *
-     * @param  mixed  $value
-     * @return \Illuminate\Database\Query\Expression
-     */
-    public function raw($value)
-    {
-    }
-
-    /**
-     * Run a select statement and return a single result.
-     *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @return mixed
-     */
-    public function selectOne($query, $bindings = [], $useReadPdo = true)
-    {
-    }
-
-    /**
      * Run a select statement against the database.
-     *
-     * @param  array   $params
-     * @param  array   $bindings
-     * @return array
-     */
-    public function select($params, $bindings = [], $useReadPdo = true)
-    {
-        return $this->run(
-            $this->addClientParams($params),
-            $bindings,
-            Closure::fromCallable([$this->connection, 'search'])
-        );
-    }
-
-    /**
-     * Run a select statement against the database and return a generator.
-     *
-     * @param  string  $query
-     * @param  array  $bindings
-     * @param  bool  $useReadPdo
-     * @return \Generator
-     */
-    public function cursor($query, $bindings = [], $useReadPdo = false)
-    {
-        $scrollTimeout = '30s';
-        $limit = $query['size'] ?? 0;
-
-        $scrollParams = array(
-            'scroll' => $scrollTimeout,
-            'size'   => 100, // Number of results per shard
-            'index'  => $query['index'],
-            'body'   => $query['body']
-        );
-
-        $results = $this->select($scrollParams);
-
-        $scrollId = $results['_scroll_id'];
-
-        $numResults = count($results['hits']['hits']);
-
-        foreach ($results['hits']['hits'] as $result) {
-            yield $result;
-        }
-
-        if (! $limit || $limit > $numResults) {
-            $limit = $limit ? $limit - $numResults : $limit;
-
-            foreach ($this->scroll($scrollId, $scrollTimeout, $limit) as $result) {
-                yield $result;
-            }
-        }
-    }
-
-    /**
-     * Run a select statement against the database using an Elasticsearch scroll cursor.
-     *
-     * @param  string $scrollId
-     * @param  string $scrollTimeout
-     * @param  int $limit
-     * @return \Generator
-     */
-    public function scroll(string $scrollId, string $scrollTimeout = '30s', int $limit = 0)
-    {
-        $numResults = 0;
-
-        // Loop until the scroll 'cursors' are exhausted or we have enough results
-        while (! $limit || $numResults < $limit) {
-            // Execute a Scroll request
-            $results = $this->connection->scroll(array(
-                'scroll_id' => $scrollId,
-                'scroll'    => $scrollTimeout,
-            ));
-
-            // Get new scroll ID in case it's changed
-            $scrollId = $results['_scroll_id'];
-
-            // Break if no results
-            if (empty($results['hits']['hits'])) {
-                break;
-            }
-
-            foreach ($results['hits']['hits'] as $result) {
-                $numResults++;
-
-                if ($limit && $numResults > $limit) {
-                    break;
-                }
-
-                yield $result;
-            }
-        }
-    }
-
-    /**
-     * Run an insert statement against the database.
      *
      * @param array $params
      * @param array $bindings
@@ -305,73 +255,28 @@ class Connection extends BaseConnection
     }
 
     /**
-     * Run an update statement against the database.
+     * Log a query in the connection's query log.
      *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @return array
+     * @param string     $query
+     * @param array      $bindings
+     * @param float|null $time
+     *
+     * @return void
      */
-    public function update($query, $bindings = [])
+    public function logQuery($query, $bindings, $time = null)
     {
-        return $this->run(
-            $query,
-            $bindings,
-            Closure::fromCallable([$this->connection, 'index'])
-        );
-    }
+        $this->event(new QueryExecuted(json_encode($query), $bindings, $time, $this));
 
-    /**
-     * Run a delete statement against the database.
-     *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @return array
-     */
-    public function delete($query, $bindings = [])
-    {
-        return $this->run(
-            $query,
-            $bindings,
-            Closure::fromCallable([$this->connection, 'delete'])
-        );
-    }
-
-    /**
-     * Execute an SQL statement and return the boolean result.
-     *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @return bool
-     */
-    public function statement($query, $bindings = [])
-    {
-    }
-
-    /**
-     * Run an SQL statement and get the number of rows affected.
-     *
-     * @param  string  $query
-     * @param  array   $bindings
-     * @return int
-     */
-    public function affectingStatement($query, $bindings = [])
-    {
-    }
-
-    /**
-     * Run a raw, unprepared query against the PDO connection.
-     *
-     * @param  string  $query
-     * @return bool
-     */
-    public function unprepared($query)
-    {
+        if ($this->loggingQueries) {
+            $this->queryLog[] = compact('query', 'bindings', 'time');
+        }
     }
 
     /**
      * Prepare the query bindings for execution.
      *
-     * @param  array  $bindings
+     * @param array $bindings
+     *
      * @return array
      */
     public function prepareBindings(array $bindings)
@@ -380,55 +285,27 @@ class Connection extends BaseConnection
     }
 
     /**
-     * Run a search query.
+     * Execute the given callback in "dry run" mode.
      *
-     * @param  array     $query
-     * @param  array     $bindings
-     * @param  \Closure  $callback
-     * @return mixed
+     * @param \Closure $callback
      *
-     * @throws \DesignMyNight\Elasticsearch\Exceptions\QueryException
+     * @return array
      */
-    protected function runQueryCallback($query, $bindings, Closure $callback)
+    public function pretend(Closure $callback)
     {
-        try {
-            $result = $callback($query, $bindings);
-        } catch (Exception $e) {
-            throw new QueryException($query, null, $e);
-        }
-
-        return $result;
+        //
     }
 
     /**
-     * Execute a Closure within a transaction.
+     * Get a new raw query expression.
      *
-     * @param  \Closure  $callback
-     * @param  int  $attempts
-     * @return mixed
+     * @param mixed $value
      *
-     * @throws \Throwable
+     * @return \Illuminate\Database\Query\Expression
      */
-    public function transaction(Closure $callback, $attempts = 1)
+    public function raw($value)
     {
-    }
-
-    /**
-     * Start a new database transaction.
-     *
-     * @return void
-     */
-    public function beginTransaction()
-    {
-    }
-
-    /**
-     * Commit the active database transaction.
-     *
-     * @return void
-     */
-    public function commit()
-    {
+        //
     }
 
     /**
@@ -438,39 +315,108 @@ class Connection extends BaseConnection
      */
     public function rollBack($toLevel = null)
     {
+        //
     }
 
     /**
-     * Get the number of active transactions.
+     * Run a select statement against the database using an Elasticsearch scroll cursor.
      *
-     * @return int
+     * @param string $scrollId
+     * @param string $scrollTimeout
+     * @param int    $limit
+     *
+     * @return \Generator
      */
-    public function transactionLevel()
+    public function scroll(string $scrollId, string $scrollTimeout = '30s', int $limit = 0)
     {
+        $numResults = 0;
+
+        // Loop until the scroll 'cursors' are exhausted or we have enough results
+        while (!$limit || $numResults < $limit) {
+            // Execute a Scroll request
+            $results = $this->connection->scroll([
+                'scroll_id' => $scrollId,
+                'scroll' => $scrollTimeout,
+            ]);
+
+            // Get new scroll ID in case it's changed
+            $scrollId = $results['_scroll_id'];
+
+            // Break if no results
+            if (empty($results['hits']['hits'])) {
+                break;
+            }
+
+            foreach ($results['hits']['hits'] as $result) {
+                $numResults++;
+
+                if ($limit && $numResults > $limit) {
+                    break;
+                }
+
+                yield $result;
+            }
+        }
     }
 
     /**
-     * Execute the given callback in "dry run" mode.
+     * Run an update statement against the database.
      *
-     * @param  \Closure  $callback
      * @return array
      */
-    public function pretend(Closure $callback)
+    public function select($params, $bindings = [], $useReadPdo = true)
     {
+        return $this->run(
+            $this->addClientParams($params),
+            $bindings,
+            Closure::fromCallable([$this->connection, 'search'])
+        );
     }
 
     /**
-     * Get the timeout for the entire Elasticsearch request
-     * @return float
+     * Run a select statement and return a single result.
+     *
+     * @param string $query
+     * @param array  $bindings
+     *
+     * @return mixed
      */
-    public function getRequestTimeout(): float
+    public function selectOne($query, $bindings = [], $useReadPdo = true)
     {
-        return $this->requestTimeout;
+        //
+    }
+
+    /**
+     * Get a new query builder instance.
+     *
+     * @return
+     */
+    public function query()
+    {
+        return new QueryBuilder(
+            $this, $this->getQueryGrammar(), $this->getPostProcessor()
+        );
+    }
+
+    /**
+     * Set the table prefix in use by the connection.
+     *
+     * @param string $prefix
+     *
+     * @return void
+     */
+    public function setIndexSuffix($suffix)
+    {
+        $this->indexSuffix = $suffix;
+
+        $this->getQueryGrammar()->setIndexSuffix($suffix);
     }
 
     /**
      * Get the timeout for the entire Elasticsearch request
+     *
      * @param float $requestTimeout seconds
+     *
      * @return self
      */
     public function setRequestTimeout(float $requestTimeout): self
@@ -481,8 +427,101 @@ class Connection extends BaseConnection
     }
 
     /**
+     * Execute an SQL statement and return the boolean result.
+     *
+     * @param string $query
+     * @param array  $bindings
+     *
+     * @return bool
+     */
+    public function statement($query, $bindings = [], Blueprint $blueprint = null)
+    {
+        //
+    }
+
+    /**
+     * Execute a Closure within a transaction.
+     *
+     * @param \Closure $callback
+     * @param int      $attempts
+     *
+     * @return mixed
+     *
+     * @throws \Throwable
+     */
+    public function transaction(Closure $callback, $attempts = 1)
+    {
+        //
+    }
+
+    /**
+     * Get the number of active transactions.
+     *
+     * @return int
+     */
+    public function transactionLevel()
+    {
+        //
+    }
+
+    /**
+     * Run a raw, unprepared query against the PDO connection.
+     *
+     * @param string $query
+     *
+     * @return bool
+     */
+    public function unprepared($query)
+    {
+        //
+    }
+
+    /**
+     * Run an update statement against the database.
+     *
+     * @param string $query
+     * @param array  $bindings
+     *
+     * @return array
+     */
+    public function update($query, $bindings = [])
+    {
+        $updateMethod = isset($query['body']['query']) ? 'updateByQuery' : 'update';
+        return $this->run(
+            $query,
+            $bindings,
+            Closure::fromCallable([$this->connection, $updateMethod])
+        );
+    }
+
+    /**
+     * @param string $index
+     * @param array  $body
+     */
+    public function updateIndex(string $index, array $body): void
+    {
+        $this->indices()->putMapping(compact('index', 'body'));
+    }
+
+    /**
+     * Set the table prefix and return the grammar.
+     *
+     * @param \Illuminate\Database\Grammar $grammar
+     *
+     * @return \Illuminate\Database\Grammar
+     */
+    public function withIndexSuffix(BaseGrammar $grammar)
+    {
+        $grammar->setIndexSuffix($this->indexSuffix);
+
+        return $grammar;
+    }
+
+    /**
      * Add client-specific parameters to the request params
+     *
      * @param array $params
+     *
      * @return array
      */
     protected function addClientParams(array $params): array
@@ -495,15 +534,91 @@ class Connection extends BaseConnection
     }
 
     /**
-     * Dynamically pass methods to the connection.
+     * Create a new Elasticsearch connection.
      *
-     * @param string $method
-     * @param array  $parameters
+     * @param array $hosts
+     * @param array $config
+     *
+     * @return \Elasticsearch\Client
+     */
+    protected function createConnection($hosts, array $config, array $options)
+    {
+        // apply config to each host
+        $hosts = array_map(function ($host) use ($config) {
+            $port = !empty($config['port']) ? $config['port'] : 9200;
+
+            $scheme = !empty($config['scheme']) ? $config['scheme'] : 'http';
+
+            // force https for port 443
+            $scheme = (int) $port === 443 ? 'https' : $scheme;
+
+            return [
+                'host' => $host,
+                'port' => $port,
+                'scheme' => $scheme,
+                'user' => !empty($config['username']) ? $config['username'] : null,
+                'pass' => !empty($config['password']) ? $config['password'] : null,
+            ];
+        }, $hosts);
+
+        $clientBuilder = ClientBuilder::create()
+            ->setHosts($hosts);
+
+        $elasticConfig = config('elasticsearch.connections.' . config('elasticsearch.defaultConnection', 'default'), []);
+        // Set additional client configuration
+        foreach ($this->configMappings as $key => $method) {
+            $value = Arr::get($elasticConfig, $key);
+            if (is_array($value)) {
+                foreach ($value as $vItem) {
+                    $clientBuilder->$method($vItem);
+                }
+            } elseif ($value !== null) {
+                $clientBuilder->$method($value);
+            }
+        }
+
+        return $clientBuilder->build();
+    }
+
+    /**
+     * Get the default post processor instance.
+     *
+     * @return Processor
+     */
+    protected function getDefaultPostProcessor()
+    {
+        return new QueryProcessor();
+    }
+
+    /**
+     * Get the default query grammar instance.
+     *
+     * @return \Illuminate\Database\Query\Grammars\Grammar
+     */
+    protected function getDefaultQueryGrammar()
+    {
+        return $this->withIndexSuffix(new QueryGrammar);
+    }
+
+    /**
+     * Run a search query.
+     *
+     * @param array    $query
+     * @param array    $bindings
+     * @param \Closure $callback
      *
      * @return mixed
+     *
+     * @throws QueryException
      */
-    public function __call($method, $parameters)
+    protected function runQueryCallback($query, $bindings, Closure $callback)
     {
-        return call_user_func_array([$this->connection, $method], $parameters);
+        try {
+            $result = $callback($query, $bindings);
+        } catch (\Exception $e) {
+            throw new QueryException($query, $bindings, $e);
+        }
+
+        return $result;
     }
 }
